backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings==2.1.0

# AI and ML libraries
openai>=1.6.1,<2.0.0
google-generativeai==0.3.2
chromadb==0.4.18
pymupdf==1.23.8
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.5

# Web search
google-search-results==2.4.2

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27 